manage.py (pre-commit): Build geopackage template and docs
