--- metadata_reports/networks.csv
@@ -1,15 +1,0 @@
-Name,Type,Not Null,Auto Increment,Title,Description,Enum,Reference Table,Reference Column
-id,INTEGER,FALSE,TRUE,,,,,
-ofds_id,TEXT,FALSE,FALSE,Identifier,"A universally unique identifier for this network, as defined by [RFC 4122](https://datatracker.ietf.org/doc/html/rfc4122). For more information, see the [identifier reference](https://open-fibre-data-standard.readthedocs.io/en/latest/reference/identifiers.html).",,,
-name,TEXT,FALSE,FALSE,Network name,A name for this network.,,,
-website,TEXT,FALSE,FALSE,Website,The URL of the website for this network.,,,
-publisher__name,TEXT,FALSE,FALSE,Publisher Name,The publisher of this network. A name for this organisation.,,,
-publisher__identifier__id,TEXT,FALSE,FALSE,Publisher Organisation identifier Identifier,The publisher of this network. An identifier for this organisation. The identifier assigned to the organisation in the register identified in `.scheme`.,,,
-publisher__identifier__scheme,TEXT,FALSE,FALSE,Publisher Organisation identifier Scheme,"The publisher of this network. An identifier for this organisation. The register from which the identifier in `.id` is drawn, from the open [organisationIdentifierScheme codelist](https://open-fibre-data-standard.readthedocs.io/en/latest/reference/codelists.html#organisationidentifierscheme).",,,
-publisher__identifier__legalName,TEXT,FALSE,FALSE,Publisher Organisation identifier Legal name,The publisher of this network. An identifier for this organisation. The legally registered name of the organisation,,,
-publisher__identifier__uri,TEXT,FALSE,FALSE,Publisher Organisation identifier URI,"The publisher of this network. An identifier for this organisation. A canonical URI for this identifier, such as those provided by Open Corporates. Do not use this field to provide the website of the organisation: instead, use `Organisation.website`.",,,
-publicationDate,TEXT,FALSE,FALSE,Publication date,The date when this network was published.,,,
-collectionDate,TEXT,FALSE,FALSE,Collection date,"The date when the location data was collected. If a dataset was produced by digitising a map, the date that the data for the map was collected.",,,
-accuracy,REAL,FALSE,FALSE,Accuracy,"The horizontal uncertainty, in metres, of the coordinates in this dataset relative to the datum of the coordinate reference system specified in `crs`. Further details about the accuracy of coordinates in this dataset can be provided in `.accuracyDetails`.",,,
-accuracyDetails,TEXT,FALSE,FALSE,Accuracy details,"Further details about the accuracy specified in `accuracy`. For example, the confidence level of the accuracy measurement, the methodology used to calculate the accuracy, or a local classification of accuracy.",,,
-language,TEXT,FALSE,FALSE,Language,"The default language of this network,from the open [language codelist](https://open-fibre-data-standard.readthedocs.io/en/latest/reference/codelists.html#language). A [BCP47 language tag](https://www.w3.org/International/articles/language-tags/) is allowed, if there is a user need for the additional information.",,,
