+++ schema/geopackage/table_definitions/codelist_closed_deployment.csv
