+++ schema/geopackage/table_definitions/codelist_closed_transmissionMedium.csv
