+++ schema/geopackage/table_definitions/codelist_open_nodeTechnologies.csv
