+++ schema/geopackage/table_definitions/codelist_open_spanTechnologies.csv
