+++ schema/geopackage/table_definitions/codelist_tables.csv
@@ -1,0 +1,11 @@
+`codelist_open_organisationIdentifierScheme`,[organisationIdentifierScheme](../codelists.md#organisationidentifierscheme)
+`codelist_open_language`,[language](../codelists.md#language)
+`codelist_open_contractType`,[contractType](../codelists.md#contracttype)
+`codelist_open_mediaType`,[mediaType](../codelists.md#mediatype)
+`codelist_open_nodeType`,[nodeType](../codelists.md#nodetype)
+`codelist_open_nodeTechnologies`,[nodeTechnologies](../codelists.md#nodetechnologies)
+`codelist_closed_transmissionMedium`,[transmissionMedium](../codelists.md#transmissionmedium)
+`codelist_closed_deployment`,[deployment](../codelists.md#deployment)
+`codelist_open_spanTechnologies`,[spanTechnologies](../codelists.md#spantechnologies)
+`codelist_closed_country`,[country](../codelists.md#country)
+`codelist_open_organisationRole`,[organisationRole](../codelists.md#organisationrole)
