+++ schema/geopackage/table_definitions/contracts_documents.csv
@@ -3,6 +3,6 @@ id,INTEGER,FALSE,TRUE,,,,,
 title,TEXT,FALSE,FALSE,Title,A title for this document.,,,
 description,TEXT,FALSE,FALSE,Description,"A description of this document. Descriptions should not exceed 250 words. If the document is not accessible online, the description should describe how to access a copy of the document.",,,
 url,TEXT,FALSE,FALSE,URL,A web address for accessing this document.,,,
-format,TEXT,FALSE,FALSE,Format,"The format of this document, from the open [mediaType codelist](https://open-fibre-data-standard.readthedocs.io/en/latest/reference/codelists.html#mediatype).",,,
-network_id,TEXT,FALSE,FALSE,Network ID,,,,
-contract_id,TEXT,FALSE,FALSE,Contract ID,,,,
+format,INTEGER,FALSE,FALSE,Format,"The format of this document, from the open [mediaType codelist](https://open-fibre-data-standard.readthedocs.io/en/latest/reference/codelists.html#mediatype).",,codelist_open_mediaType,id
+network_id,INTEGER,FALSE,FALSE,Network ID,,,networks,id
+contract_id,INTEGER,FALSE,FALSE,Contract ID,,,contracts,id
