+++ schema/geopackage/table_definitions/nodes_internationalConnections.csv
@@ -5,5 +5,5 @@ locality,TEXT,FALSE,FALSE,Locality,"The locality. For example, Mountain View.",,
 region,TEXT,FALSE,FALSE,Region,"The region. For example, CA.",,,
 postalCode,TEXT,FALSE,FALSE,Postal code,"The postal code. For example, 94043.",,,
 country,TEXT,FALSE,FALSE,Country,"The country in which the address is physically located, from the closed [country codelist](https://open-fibre-data-standard.readthedocs.io/en/latest/reference/codelists.html#country).",closed_codelist_country.csv,,
-network_id,TEXT,FALSE,FALSE,Network ID,,,,
-node_id,TEXT,FALSE,FALSE,Nodes ID,,,,
+network_id,INTEGER,FALSE,FALSE,Network ID,,,networks,id
+node_id,INTEGER,FALSE,FALSE,Nodes ID,,,nodes,id
