+++ schema/geopackage/table_definitions/organisations.csv
@@ -1,0 +1,13 @@
+Name,Type,Not Null,Auto Increment,Title,Description,Enum,Reference Table,Reference Column
+id,INTEGER,FALSE,TRUE,,,,,
+ofds_id,TEXT,FALSE,FALSE,Local identifier,A local identifier for this organisation. The identifier must be unique within the scope of the `organisations` array.,,,
+name,TEXT,FALSE,FALSE,Name,A name for this organisation.,,,
+identifier__id,TEXT,FALSE,FALSE,Organisation identifier Identifier,An organisation identifier. The identifier assigned to the organisation in the register identified in `.scheme`.,,,
+identifier__scheme,INTEGER,FALSE,FALSE,Organisation identifier Scheme,"An organisation identifier. The register from which the identifier in `.id` is drawn, from the open [organisationIdentifierScheme codelist](https://open-fibre-data-standard.readthedocs.io/en/latest/reference/codelists.html#organisationidentifierscheme).",,codelist_open_organisationIdentifierScheme,id
+identifier__legalName,TEXT,FALSE,FALSE,Organisation identifier Legal name,An organisation identifier. The legally registered name of the organisation,,,
+identifier__uri,TEXT,FALSE,FALSE,Organisation identifier URI,"An organisation identifier. A canonical URI for this identifier, such as those provided by Open Corporates. Do not use this field to provide the website of the organisation: instead, use `Organisation.website`.",,,
+country,TEXT,FALSE,FALSE,Country,"The country in which this organisation is legally registered, from the closed [country codelist](https://open-fibre-data-standard.readthedocs.io/en/latest/reference/codelists.html#country).",closed_codelist_country.csv,,
+roleDetails,TEXT,FALSE,FALSE,Role details,Further details about this organisation's roles in the network,,,
+website,TEXT,FALSE,FALSE,Website,The URL of the website for this organisation.,,,
+logo,TEXT,FALSE,FALSE,Logo,The URL of the logo for this organisation.,,,
+network_id,INTEGER,FALSE,FALSE,Network ID,,,networks,id
