+++ schema/geopackage/table_definitions/relation_contracts_relatedPhases.csv
