+++ schema/geopackage/table_definitions/relation_nodes_networkProviders.csv
