+++ schema/geopackage/table_definitions/relation_nodes_type.csv
