+++ schema/geopackage/table_definitions/relation_organisations_roles.csv
