+++ schema/geopackage/table_definitions/relation_spans_deployment.csv
