+++ schema/geopackage/table_definitions/relation_spans_networkProviders.csv
