+++ schema/geopackage/table_definitions/spans.csv
@@ -1,0 +1,26 @@
+Name,Type,Not Null,Auto Increment,Title,Description,Enum,Reference Table,Reference Column
+id,INTEGER,FALSE,TRUE,,,,,
+geom,BLOB,FALSE,FALSE,,,,,
+ofds_id,TEXT,FALSE,FALSE,Identifier,An identifier for this span. The identifier must be unique within the scope of the `spans` array.,,,
+name,TEXT,FALSE,FALSE,Span name,A name for this span.,,,
+phase__id,TEXT,FALSE,FALSE,Phase reference Identifier,The identifier and name of the phase being referenced. Used for cross-referencing. The identifier of the phase being referenced. This must match the `.id` of a phase in the `phases` array.,,,
+phase__name,TEXT,FALSE,FALSE,Phase reference Name,The identifier and name of the phase being referenced. Used for cross-referencing. The name of the phase being referenced. This must match the `.name` of the phase in the `phases` array whose `.id` matches the `.id` of this phase reference.,,,
+status,TEXT,FALSE,FALSE,Span status,"The status of the network infrastructure for this span, from the closed [spanStatus codelist](https://open-fibre-data-standard.readthedocs.io/en/latest/reference/codelists.html#spanstatus).",closed_codelist_spanStatus.csv,,
+readyForServiceDate,TEXT,FALSE,FALSE,Ready for service date,"The date this span was ready to carry traffic. For spans with multiple network providers, the earliest of the dates at which each provider's network was ready to carry traffic on this span.",,,
+start,INTEGER,FALSE,FALSE,Start,"The identifier of one of two nodes that this span connects. It must match the `.id` of a node in the `nodes` array. If `directed == true`, it represents the starting point of the span. Otherwise, the distinction between `start` and `end` is arbitrary.",,nodes,id
+end,INTEGER,FALSE,FALSE,End,"The identifier of one of two nodes that this span connects. It must match the `.id` of a node in the `nodes` array. If `directed == true`, it represents the end point of the span. Otherwise, the distinction between `start` and `end` is arbitrary.",,nodes,id
+directed,TEXT,FALSE,FALSE,Directed,"Whether this span is directed. If `True`, then `.start` represents the start of the span and `.end` represents the end. Otherwise, `.start` and `.end` reference this span's endpoints with no direction implied.",,,
+physicalInfrastructureProvider__id,TEXT,FALSE,FALSE,Organisation reference Local identifier,The local identifier and name of the organisation being referenced. Used to cross-reference to the `organisations` array. The local identifier of the organisation being referenced. This must match the `.id` of an organisation in the `organisations` array.,,,
+physicalInfrastructureProvider__name,TEXT,FALSE,FALSE,Organisation reference Name,The local identifier and name of the organisation being referenced. Used to cross-reference to the `organisations` array. The name of the organisation being referenced. This must match the `.name` of the organisation in the `organisations` array whose `.id` matches the `.id` of this organisation reference.,,,
+supplier__id,TEXT,FALSE,FALSE,Organisation reference Local identifier,The local identifier and name of the organisation being referenced. Used to cross-reference to the `organisations` array. The local identifier of the organisation being referenced. This must match the `.id` of an organisation in the `organisations` array.,,,
+supplier__name,TEXT,FALSE,FALSE,Organisation reference Name,The local identifier and name of the organisation being referenced. Used to cross-reference to the `organisations` array. The name of the organisation being referenced. This must match the `.name` of the organisation in the `organisations` array whose `.id` matches the `.id` of this organisation reference.,,,
+deploymentDetails__description,TEXT,FALSE,FALSE,Deployment details Description,Further details about a span's deployment. A description of this span's deployment.,,,
+darkFibre,TEXT,FALSE,FALSE,Dark fibre availability,Whether access to dark fibre is available on this span.,,,
+fibreType,TEXT,FALSE,FALSE,Fibre type,"The type of fibre used in this span, from the closed [fibreType codelist](https://open-fibre-data-standard.readthedocs.io/en/latest/reference/codelists.html#fibretype). Further details of the span's fibre type can be provided in `.fibreTypeDetails`.",closed_codelist_fibreType.csv,,
+fibreTypeDetails__fibreSubtype,TEXT,FALSE,FALSE,Fibre type details Fibre subtype,"Further details about a span's fibre type. The sub-category of the fibre type. For example, G.652.B.",,,
+fibreTypeDetails__description,TEXT,FALSE,FALSE,Fibre type details Description,Further details about a span's fibre type. A description of this span's fibre type.,,,
+fibreCount,INTEGER,FALSE,FALSE,Fibre count,The number of individual optical fibres in this span.,,,
+fibreLength,REAL,FALSE,FALSE,Fibre length,"The physical length in kilometres, of fibre optic cable used in this span.",,,
+capacity,REAL,FALSE,FALSE,Capacity,"The transmission rate, or throughput, of this span, expressed in Gbit/sec (Gbps). The equipped capacity is the total capacity of the circuits (e.g. E1, DS3, STM-1 etc.) which have been activated in the network transmission equipment of the span. For spans with multiple network providers, the sum of the capacity of each provider's network on this span. Further details of this span's capacity can be provided in `.capacityDetails`.",,,
+capacityDetails__description,TEXT,FALSE,FALSE,Capacity details Description,Further details about a span's capacity. A description of this span's capacity.,,,
+network_id,INTEGER,FALSE,FALSE,Network ID,,,networks,id
